Dont change packet id when retransmitting request.
When a packet is retranmitted it should use the same ID so the server
can detect it as a retransmission, if it was the reply that was lost.

Also, dont decrement the configured retries for the request channel
so the same socket and connection can be used for multiple requests.